pyproject.toml
[project]
name = "the-economist-poke-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.13.1",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
